+++ test/cider-client-tests.el
@@ -498,3 +498,20 @@ SYMBOL is locally let-bound to the current buffer."
     (spy-on 'cider-connected-p :and-return-value nil)
     (spy-on 'clojure-expected-ns :and-return-value "clojure-expected-ns")
     (expect (cider-expected-ns "foo") :to-equal "clojure-expected-ns")))
+
+(describe "cider-load-file"
+  (it "works as expected in empty Clojure buffers"
+    (spy-on 'cider-request:load-file :and-return-value nil)
+    (with-connection-buffer "clj" b
+     (with-temp-buffer
+       (clojure-mode)
+       (setq buffer-file-name (make-temp-name "tmp.clj"))
+       (expect (lambda () (cider-load-buffer)) :not :to-throw)))))
+
+(describe "cider-interactive-eval"
+  (it "works as expected in empty Clojure buffers"
+    (spy-on 'cider-nrepl-request:eval :and-return-value nil)
+    (with-connection-buffer "clj" b
+      (with-temp-buffer
+        (clojure-mode)
+        (expect (lambda () (cider-interactive-eval "(+ 1)")) :not :to-throw)))))